requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.4
django-phonenumber-field==5.0.0
djangorestframework==3.12.2
idna==2.10
phonenumberslite==8.12.15
PyJWT==2.0.0
pyotp==2.4.1
python-decouple==3.3
pytz==2020.5
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
twilio==6.50.1
urllib3==1.26.2
